feat(#32): local mail helper (다형성 적용)
